LPS-75220 Move BoundaryCheckerUtil to where it is used.
