+++ modules/core/petra/petra-io/src/main/java/com/liferay/petra/io/unsync/UnsyncStringWriter.java
@@ -14,7 +14,6 @@
 
 package com.liferay.petra.io.unsync;
 
-import com.liferay.petra.io.internal.BoundaryCheckerUtil;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.petra.string.StringPool;
